pyproject.toml
[tool.poetry]
name = "scrat"
version = "0.1.0"
description = ""
authors = ["Lajos Santa <santa.lajos@gmail.com>"]
license = "MIT"
include = ["scrat/web/static/*"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"
Flask = "^1.1.2"
SQLAlchemy = "^1.3.22"
alembic = "^1.4.3"
logger = "^1.4"
configpp = "^0.5.0"
redis = "^3.5.3"
aioredis = "^1.3.1"
psycopg2 = "^2.8.6"
tabulate = "^0.8.7"
wows-api-async = "^0.1.0"

[tool.poetry.dev-dependencies]
invoke = "^1.5.0"
livereload = "^2.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
